Create NeggHeads
This contract defines a token called NeggHeads that is an ERC721 non-fungible token. It has a mint function that allows users to mint new NeggHeads with random face, color, eye, and mouth traits, as well as a rarity ranking. It also has a getNeggHead function that allows users to query the traits and rarity of a specific NeggHead.