Skim any UniV2 pair w/ built in "skim" function
This contract defines a constructor that takes the address of the Uniswap V2 pair ERC20 token as an input and stores it in a pair variable. The contract also defines a skim function that calls the skim function of the pair contract.

In order to use this contract, you would need to deploy it and pass the address of the Uniswap V2 pair ERC20 token as an argument to the constructor. Then, you can call the skim function of the contract from an external address to execute the skim function of the Uniswap V2 pair ERC20 token.