Test Staking v2 Code by ChatGPT
This contract allows users to stake their ERC1155 tokens by calling the stake function and passing in an array of token IDs and an array of amounts to stake. The contract then uses the safeBatchTransferFrom function from the ERC1155 contract to transfer the staked tokens from the user's account to the contract's account.

The contract also has a distributeRewards function that can be called to distribute rewards to stakers. This function calculates the total amount of staked tokens and the total amount of reward tokens, and then uses the safeTransfer function from the ERC20 contract to transfer the rewards to each staker in proportion to the amount of tokens they have staked.