Partial example code from ChatGPT for ERC1155 stake
To modify the contract to allow for ERC1155 token deposits, you will need to replace the IERC721 interface with the IERC1155 interface from the OpenZeppelin contracts library. Additionally, you will need to update the mappings and structs that use the uint256 type to use the uint256[] type to accommodate for the multiple token ID's that ERC1155 contracts can contain.